[fix][s] change resource create and package update - #39
