+++ public/index.html
@@ -28,7 +28,7 @@
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
-    <div id="ResourceEditor" data-dataset-id="test-id"></div>
+    <div id="ResourceEditor" data-api="http://127.0.0.1:5000"  data-lfs="http://localhost:9419" data-auth-token="be270cae-1c77-4853-b8c1-30b6cf5e9878" data-dataset-id="sample_1" data-organization-id="myorg"></div>
     <!--
       This HTML file is a template.
       If you open it directly in the browser, you will see an empty page.
